Fixing framebuffer black on resize
The problem was that a new framebuffer was created on resizing the viewport and inmediatly after that was sended to imgui to get the texture and draw it in the viewport window. But there was nothing to draw becuase the render update had not been ocurred until the next frame, resulting in a black screen. The solution was to check for the changes of the viewport size inside the OnUpdate method