Added a new Texture2D class
The old one will be TextureMesh2D. It's more semantic that way because that texture is directly related to a mesh and it has type (e.g diffuse, specular, etc). The new one Texture2D is more like an OpenGL texture and it create/bind/delete a buffer etc. This class has a method to set Data manually and not from an image file