Added the shadow map calculation to the TerrainScene
Added a simple window in the Editor layer to see the depth map of the shadow mapping. Added new variables to the SceneCompositionPanelProps to set the correct Orthographic projection. Added a the shader for the shadow pass and added shadow calculation to the terrain shader. The model shader wont include the shadows calculations (added would imply the shadow of the objects being cast to itselfs, which is not necessary in this context, as the scene is rendered far away)